main: Update datajam.Dockerfile

+++ mspy_vendi/domain/impressions/schemas.py
@@ -1,4 +1,5 @@
 from datetime import date
+from decimal import Decimal
 
 from mspy_vendi.core.constants import DEFAULT_SOURCE_SYSTEM
 from mspy_vendi.core.schemas import BaseSchema
@@ -6,7 +7,7 @@ from mspy_vendi.core.schemas import BaseSchema
 
 class CreateImpressionSchema(BaseSchema):
     date: date
-    total_impressions: int
+    total_impressions: Decimal
     temperature: int
     rainfall: int
     source_system: str = DEFAULT_SOURCE_SYSTEM
